pyproject.toml
[project]
name = "one-time-secrets"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flake8 (>=7.2.0,<8.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "django (>=5.2,<6.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '.venv'

[tool.isort]
line_length = 119